libtorch: restore compat python_api shim and dense-fp8 extension
build: replace torch/extension.h with libtorch headers

build: remove torch/python.h include from kerutils tensor helpers

build: avoid utils.h include collision with sgl-kernel

fix: keep FMHACutlassSM100FwdRun ABI and cast internally

compat: convert sparse prefill max_logits/lse back to log2

+++ csrc/sm100/prefill/dense/interface.h
@@ -1,11 +1,14 @@
 #pragma once
 
+#include <cstdint>
+
 #include <ATen/Tensor.h>
 
 void FMHACutlassSM100FwdRun(at::Tensor workspace_buffer, at::Tensor q, at::Tensor k, at::Tensor v,
                             at::Tensor cumulative_seqlen_q, at::Tensor cumulative_seqlen_kv,
                             at::Tensor o, at::Tensor lse,
-                            int mask_mode_code, float softmax_scale, int max_seqlen_q, int max_seqlen_kv, bool is_varlen);
+                            int64_t mask_mode_code, double softmax_scale, int64_t max_seqlen_q, int64_t max_seqlen_kv,
+                            bool is_varlen);
 
 void FMHACutlassSM100BwdRun(at::Tensor workspace_buffer, at::Tensor d_o, at::Tensor q, at::Tensor k,
                             at::Tensor v, at::Tensor o, at::Tensor lse,
